// frontend/src/types/product.type.ts
export type ProductType = {
  id: string
  name: string
  price: number
  image: string
  lightning: string
  humidity: string
  temperature: string
  height: number
  diameter: string
  url: string
  type: {
    id: string
    name: string
    url: string
  },
  countInCart?: number
  isInFavourite?: boolean
}